Remove the test emitter code from MainNoGUI.cpp file. Change back from u32* pointers to u8* pointers just to stay consistant with the other JIT cores. Complete some more of the dispatcher code, this actually now calls the code that is in the JitCacheBlock and branches correctly. Also fix a bunch of emitter functions that were implemented by Calc83 but they were slightly off. Also have both the immediate and register versions of those instructions.

+++ Source/Core/Core/Src/PowerPC/JitArm32/JitAsm.h
@@ -40,7 +40,7 @@ private:
 	void GenerateCommon();
 
 public:
-	const u32 *enterCode;
+	const u8 *enterCode;
 
 	const u8 *dispatcher;
 	const u8 *dispatcherNoCheck;
